refactored all emit functions out of mod.rs for clarity for extruder
