+++ server/src/machines/extruder1/mock_emit.rs
@@ -1,0 +1,1 @@
+// imitates the api behaviours of an ExtruderV2
